feat: Move useColumns to own file and update comment
